feat: Create Mandelbrot set image.
Created methods map, which maps a real coordinate to a complex plane coordinate, identifySet, which checks to see if a complex point is in the Mandelbrot set, and setRGBColor which creates RGB values based on the iterations it took to determine Mandelbrot set membership. Used all of these methods together to create a Mandelbrot set image.